+++ bin/cli/src/validate/mod.rs
@@ -20,13 +20,13 @@ use crate::validate::proving::{create_proving_args, Task};
 use alloy::primitives::{Address, FixedBytes, B256};
 use anyhow::{anyhow, bail, Context};
 use kailua_build::KAILUA_FPVM_ID;
-use kailua_client::args::parse_address;
 use kailua_client::boundless::BoundlessArgs;
 use kailua_client::proof::{proof_file_name, read_proof_file};
 use kailua_common::config::config_hash;
 use kailua_common::journal::ProofJournal;
 use kailua_common::precondition::PreconditionValidationData;
 use kailua_host::channel::AsyncChannel;
+use kailua_sync::args::parse_address;
 use kailua_sync::args::SyncArgs;
 use kailua_sync::provider::optimism::fetch_rollup_config;
 use kailua_sync::telemetry::TelemetryArgs;
